add custom layout able to pull data from custom layout

+++ app/src/main/java/com/example/sprintchallenge3/CreateItem.kt
@@ -1,3 +1,3 @@
 package com.example.sprintchallenge3
 
-class CreateItem (val name: String, val rating: Float)
+class CreateItem (val name: String, val rating: Int)
